pyproject.toml
[project]
name = "gestalt"
version = "0.1.0"
description = "Guardrail-first Gestalt: parts compatibility UI and CrewAI crew"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "crewai[tools]>=0.100.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
gestalt-web = "app:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "agents", "crew", "compatibility_checker"]

[dependency-groups]
dev = ["pytest>=8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
